+++ src/config/mod.rs
@@ -1,1 +1,2 @@
+pub mod fragment;
 pub mod merge;