+++ stackable-operator-derive/src/lib.rs
@@ -1,57 +1,7 @@
-use darling::{
-    ast::{Data, Fields},
-    FromDeriveInput, FromField, FromMeta, FromVariant,
-};
-use proc_macro2::{Ident, Span, TokenStream};
-use quote::{format_ident, quote};
-use syn::{parse_macro_input, parse_quote, Generics, Index, Path, WherePredicate};
+use syn::parse_macro_input;
 
-#[derive(FromMeta)]
-struct PathOverrides {
-    #[darling(default = "PathOverrides::default_merge")]
-    merge: Path,
-}
-impl Default for PathOverrides {
-    fn default() -> Self {
-        Self {
-            merge: Self::default_merge(),
-        }
-    }
-}
-impl PathOverrides {
-    fn default_merge() -> Path {
-        parse_quote!(::stackable_operator::config::merge)
-    }
-}
-
-#[derive(FromDeriveInput)]
-#[darling(attributes(merge))]
-struct MergeInput {
-    ident: Ident,
-    generics: Generics,
-    data: Data<MergeVariant, MergeField>,
-    #[darling(default)]
-    path_overrides: PathOverrides,
-    #[darling(default)]
-    bounds: Option<Vec<WherePredicate>>,
-}
-
-#[derive(FromField)]
-struct MergeField {
-    ident: Option<Ident>,
-}
-
-#[derive(FromVariant)]
-struct MergeVariant {
-    ident: Ident,
-    fields: Fields<MergeField>,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-enum InputType {
-    Struct,
-    Enum,
-}
+mod fragment;
+mod merge;
 
 /// Derives [`Merge`](trait.Merge.html) for a given struct or enum, by merging each field individually.
 ///
@@ -59,7 +9,7 @@ enum InputType {
 ///
 /// # Supported attributes
 ///
-/// ## `#[merge(bounds = "...")]`
+/// ## `#[merge(bound = "...")]`
 ///
 /// This attribute can be used to specify additional `where` clauses on the derived trait implementation.
 /// Bounds specified on the struct itself are automatically inherited for the generated implementation, and
@@ -70,7 +20,7 @@ enum InputType {
 /// ```
 /// # use stackable_operator::config::merge::Merge;
 /// #[derive(Merge)]
-/// #[merge(bounds = "T: Merge")]
+/// #[merge(bound = "T: Merge")]
 /// struct Wrapper<T> where T: Clone {
 ///     inner: T,
 /// }
@@ -86,7 +36,7 @@ enum InputType {
 /// impl<T> Merge for Wrapper<T>
 /// where
 ///     T: Clone, // this clause was inherited from the struct
-///     T: Merge, // this clause was specified using #[merge(bounds)]
+///     T: Merge, // this clause was specified using #[merge(bound)]
 /// {
 ///     fn merge(&mut self, defaults: &Self) {
 ///         self.inner.merge(&defaults.inner);
@@ -100,127 +50,94 @@ enum InputType {
 /// or the `stackable_operator` crate is renamed.
 #[proc_macro_derive(Merge, attributes(merge))]
 pub fn derive_merge(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    let MergeInput {
-        ident,
-        mut generics,
-        data,
-        path_overrides: PathOverrides { merge: merge_mod },
-        bounds,
-    } = match MergeInput::from_derive_input(&parse_macro_input!(input)) {
-        Ok(input) => input,
-        Err(err) => return err.write_errors().into(),
-    };
-
-    let (ty, variants) = match data {
-        // Structs are almost single-variant enums, so we can reuse most of the same matching code for both cases
-        Data::Struct(fields) => (
-            InputType::Struct,
-            vec![MergeVariant {
-                ident: Ident::new("__placeholder", Span::call_site()),
-                fields,
-            }],
-        ),
-        Data::Enum(variants) => (InputType::Enum, variants),
-    };
-    let merge_variants = variants
-        .into_iter()
-        .map(
-            |MergeVariant {
-                 ident: variant_ident,
-                 fields,
-             }| {
-                let constructor: Path = match ty {
-                    InputType::Struct => parse_quote! {#ident},
-                    InputType::Enum => parse_quote! {#ident::#variant_ident},
-                };
-                let self_ident = format_ident!("self");
-                let defaults_ident = format_ident!("defaults");
-                let field_idents = fields.iter().map(|f| f.ident.as_ref());
-                let self_fields =
-                    map_fields_to_prefixed_vars(&constructor, field_idents.clone(), &self_ident);
-                let defaults_fields =
-                    map_fields_to_prefixed_vars(&constructor, field_idents, &defaults_ident);
-                let body = fields
-                    .into_iter()
-                    .enumerate()
-                    .map(|(field_index, field)| {
-                        let field_ident = field.ident.as_ref().ok_or(field_index);
-                        let self_field = prefix_ident(field_ident, &self_ident);
-                        let default_field = prefix_ident(field_ident, &defaults_ident);
-                        quote! {
-                            #merge_mod::Merge::merge(#self_field, #default_field);
-                        }
-                    })
-                    .collect::<TokenStream>();
-
-                let pattern = match ty {
-                    InputType::Struct => quote! {(#self_fields, #defaults_fields)},
-                    InputType::Enum => quote! {(Some(#self_fields), Some(#defaults_fields))},
-                };
-                quote! {
-                    #pattern => {#body},
-                }
-            },
-        )
-        .collect::<TokenStream>();
-
-    if let Some(bounds) = bounds {
-        let where_clause = generics.make_where_clause();
-        where_clause.predicates.extend(bounds);
-    }
-    let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
-    let ty_toks = match ty {
-        InputType::Struct => quote! { #ident #ty_generics },
-        // Enums need some way to indicate that we want to keep the same variant, in our case we use
-        // Option::None to signal this
-        InputType::Enum => quote! { Option<#ident #ty_generics> },
-    };
-    let fallback_variants = match ty {
-        InputType::Struct => quote! {},
-        InputType::Enum => quote! {
-            // self is None => inherit everything from defaults
-            (this @ None, defaults) => *this = <Self as ::std::clone::Clone>::clone(defaults),
-            // self is Some but mismatches defaults, discard defaults
-            (Some(_), _) => {}
-        },
-    };
-    quote! {
-        impl #impl_generics #merge_mod::Merge for #ty_toks #where_clause {
-            fn merge(&mut self, defaults: &Self) {
-                match (self, defaults) {
-                    #merge_variants
-                    #fallback_variants
-                }
-            }
-        }
-    }
-    .into()
+    merge::derive(parse_macro_input!(input)).into()
 }
 
-fn map_fields_to_prefixed_vars<'a>(
-    constructor: &Path,
-    fields: impl IntoIterator<Item = Option<&'a Ident>>,
-    prefix: &Ident,
-) -> TokenStream {
-    let fields = fields
-        .into_iter()
-        .enumerate()
-        .map(|(index, field)| {
-            let prefixed = prefix_ident(field.ok_or(index), prefix);
-            if let Some(field) = field {
-                quote! { #field: #prefixed, }
-            } else {
-                let index = Index::from(index);
-                quote! { #index: #prefixed, }
-            }
-        })
-        .collect::<TokenStream>();
-    quote! { #constructor { #fields } }
-}
-
-fn prefix_ident(ident: Result<&Ident, usize>, prefix: &Ident) -> Ident {
-    match ident {
-        Ok(ident) => format_ident!("{prefix}_{ident}"),
-        Err(index) => format_ident!("{prefix}_{index}"),
-    }
+/// Creates a [fragment type](index.html) for the given type, and implements [`FromFragment`](trait.FromFragment.html).
+///
+/// This macro implements "deep optionality", meaning that each field is replaced by its `Fragment` variant. For example, this type:
+///
+/// ```
+/// #[derive(Fragment)]
+/// struct Foo {
+///     atomic: u32,
+///     nested: Bar,
+/// }
+/// ```
+///
+/// would generate the following output:
+///
+/// ```
+/// struct FooFragment {
+///     atomic: u32::Fragment, // = Option<u32>
+///     nested: Bar::Fragment, // = BarFragment, assuming that Bar is also #[derive(Fragment)]
+/// }
+///
+/// impl FromFragment for Foo {
+///     // no need for Option since the value is already deeply optional
+///     type Fragment = FooFragment;
+///     type RequiredFragment = FooFragment;
+///
+///     // snipped support code
+/// }
+/// ```
+///
+/// # Supported Attributes
+///
+/// ## `#[fragment_attrs(...)]`
+///
+/// This attribute can be used to forward attributes to the generated fragment. For example, `#[fragment_attrs(derive(Default))]` derives [`Default`] for the fragment type.
+///
+/// This can be specified on both the struct itself and the field, and will be forwarded to the corresponding location on the generated fragment.
+///
+/// ## `#[fragment(bound = "...")]`
+///
+/// This attribute can be used to specify additional `where` clauses on the derived fragment and trait implementation. Bounds specified on the struct itself
+/// are automatically inherited for the generated implementation, and do not need to be repeated here.
+///
+/// ## `#[fragment(path_overrides(fragment = "..."))]`
+///
+/// This attribute can be used to override the path to the module containing the [`FromFragment`](trait.FromFragment.html) trait, if it is reexported
+/// or the `stackable_operator` crate is renamed.
+///
+/// # Generics
+///
+/// This macro supports generic types, but there are some caveats to be aware of.
+///
+/// ## Fragment macro bounds
+///
+/// The `Fragment` macro does not automatically insert any type bounds, they must be specified manually. Typically, this means adding the attribute
+/// `#[fragment(bound = "T: FromFragment")]` to the type.
+///
+/// ## Interactions with other derive macros
+///
+/// Arbitrary other macros can be specified for the generated fragment type. However, many macros automatically add the bound `where T: Trait`. This assumption makes sense
+/// for most types (with the assumption that each type is used "directly" in the struct definition).
+///
+/// However, fragment types use `T::Fragment` instead, and the generated bounds must be overridden to reflect this. The exact convention used to do this is going to vary
+/// between macros, but typically they will take an attribute such as `#[trait(bound = "T::Fragment: Trait")]`.
+///
+/// ### `std` traits
+///
+/// `std`'s built-in derive macros (such as [`Default`]) do not take any configuration. Instead, use [`Derivative`](https://mcarton.github.io/rust-derivative/latest/index.html),
+/// which supports custom bounds using the attribute convention `#[derivative(Trait(bound = "T::Fragment: Trait"))]`.
+///
+/// ### `serde`
+///
+/// Serde uses the `#[serde]` attribute to configure both `Serialize` and `Deserialize`. However, bounds must be configured separately for the two. Hence, the correct
+/// bound would be:
+///
+/// ```
+/// #[serde(bound(
+///     serialize = "T::Fragment: Serialize",
+///     deserialize = "T::Fragment: Deserialize<'de>",
+/// ))]
+/// ```
+///
+/// # Enums
+///
+/// Enums are not currently supported.
+#[proc_macro_derive(Fragment, attributes(fragment, fragment_attrs))]
+pub fn derive_fragment(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    fragment::derive(parse_macro_input!(input)).into()
 }